refactor: useConsentValueをconsentClient.tsから分離し、E2Eテストのセレクタを修正
- useConsentValue フックを lib/analytics/useConsentValue.ts に分離し、consentClient.ts を React に依存しない純粋なユーティリティに変更
- tests/e2e/analytics.spec.ts のセレクタを実装に合わせて修正（'すべて許可'、バナー本文テキスト）
